sentinel 2 to load_ard + update notebook
